+++ ui/app/src/CHANGELOG.md
@@ -32,6 +32,9 @@
   * Removed `useQuickCreateListResource` hook.
   * Removed `useSystemVersionResource` hook.
   * Removed `useResolveWhenNoNullResource` hook.
+  * Renamed `useDetailedItem` to `useContentItem`.
+  * Removed `useDetailedItems` hook.
+  * Renamed `useFetchSandboxItem` to `RenameFetchContentItem`.
 * Upgrade to the latest version to date of the following libraries:
   * @mui/icons-material
   * @mui/lab